Stabilization
I am making this fork as I still do not have permissions to publish this branch. Here is what is done so far:
The stabilization code should be in working order hopefully. However, we may need to invert some signs when telling the robot where to go as I think it may end up going in the wrong direction as it lies currently; need to actually test it to find out. Additionally, and most importantly, the actual way the stabilization is being called in HarvController is terrible at the moment, and I believe may not work at all (or cause the robot to become unresponsive) as it is is simply being called in the constructor. The call should be moved to the Robot.java class, but I do not know where to put it, hence the placement in the HarvController constructor.